fix:check comments
- donate button color change
- donate button in corousal instead of "explore work"
- welcome section above stats
- Stats section updated
- project donation images of same size
- navbar text and links fix
